[irep2] Make sideeffect_data::allockind a strongly-typed enum class
allockind was a plain enum whose members leaked into sideeffect_data's
scope (and through inheritance into sideeffect2t's), allowing implicit
integer promotion and silent comparison against unrelated values. The
enum is a discriminator picked from a small fixed list; there is no
arithmetic context where it should convert to int.

Promote it to enum class and qualify the affected call sites
(sideeffect2t::malloc -> sideeffect2t::allockind::malloc, and similarly
for the other 14 enumerators across symex, goto-programs, value-set,
the C frontend, and the migrate bridge). The do_type_crc overload that
hashed the enum through boost::hash<uint8_t> now goes through an
explicit static_cast<uint8_t>. operator< (used by do_type_lt) is
provided by enum class so the ordering overload is unchanged.

+++ src/irep2/templates/irep2_template_utils.cpp
@@ -461,7 +461,7 @@ void do_type_hash(const unsigned int &theval, crypto_hash &hash)
 
 size_t do_type_crc(const sideeffect_data::allockind &theval)
 {
-  return boost::hash<uint8_t>()(theval);
+  return boost::hash<uint8_t>()(static_cast<uint8_t>(theval));
 }
 
 void do_type_hash(const sideeffect_data::allockind &theval, crypto_hash &hash)
